requirements.txt
Flask
python-dotenv
groq
nltk
faiss-cpu
sentence-transformers
scikit-learn
rouge-score
